creating base type `SnowflakeStoreConfig` to fix circular imports
